Extract driver pack selection state from the deploy shell
