+++ src/Foundry.Deploy/ViewModels/MainWindowViewModel.cs
@@ -1,7 +1,6 @@
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Diagnostics;
-using System.Globalization;
 using System.IO;
 using System.Reflection;
 using System.Windows;
@@ -29,12 +28,6 @@ namespace Foundry.Deploy.ViewModels;
 
 public partial class MainWindowViewModel : ObservableObject, IDisposable
 {
-    private const string NoneDriverPackOptionKey = "none";
-    private const string MicrosoftUpdateCatalogDriverPackOptionKey = "microsoft-update-catalog";
-    private const string DellDriverPackOptionKey = "oem:dell";
-    private const string LenovoDriverPackOptionKey = "oem:lenovo";
-    private const string HpDriverPackOptionKey = "oem:hp";
-    private const string MicrosoftOemDriverPackOptionKey = "oem:microsoft";
     private const string DeploymentModeEnvironmentVariable = "FOUNDRY_DEPLOYMENT_MODE";
     private const string CacheVolumeLabel = "Foundry Cache";
     private const string CacheMarkerFolderName = "Foundry Cache";
@@ -50,7 +43,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     private readonly IHardwareProfileService _hardwareProfileService;
     private readonly IOfflineWindowsComputerNameService _offlineWindowsComputerNameService;
     private readonly ITargetDiskService _targetDiskService;
-    private readonly IDriverPackSelectionService _driverPackSelectionService;
     private readonly ILogger<MainWindowViewModel> _logger;
     private readonly Dispatcher _dispatcher;
     private readonly DeploymentMode _resolvedDeploymentMode;
@@ -59,8 +51,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     private DeployMachineNamingSettings _machineNamingConfiguration = new();
     private string _lockedComputerNamePrefix = string.Empty;
     private bool _isApplyingManagedComputerName;
-    private bool _isUpdatingDriverPackOptionSelection;
-    private bool _hasUserSelectedDriverPackOption;
     private bool _isUpdatingFirmwareOptionSelection;
     private bool _hasUserSelectedFirmwareOption;
     private bool _firmwareUpdatesPreference = true;
@@ -110,31 +100,8 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     private TargetDiskInfo? selectedTargetDisk;
 
     [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(DriverPackModeDisplay))]
-    [NotifyPropertyChangedFor(nameof(IsOemDriverSourceSelected))]
-    [NotifyPropertyChangedFor(nameof(IsDriverPackModelSelectionEnabled))]
-    [NotifyPropertyChangedFor(nameof(IsDriverPackVersionSelectionEnabled))]
-    [NotifyPropertyChangedFor(nameof(SelectedDriverPackSelectionDisplay))]
-    [NotifyCanExecuteChangedFor(nameof(StartDeploymentCommand))]
-    private DriverPackOptionItem? selectedDriverPackOption;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(SelectedDriverPackSelectionDisplay))]
-    [NotifyCanExecuteChangedFor(nameof(StartDeploymentCommand))]
-    private string selectedDriverPackModel = string.Empty;
-
-    [ObservableProperty]
-    [NotifyPropertyChangedFor(nameof(SelectedDriverPackSelectionDisplay))]
-    [NotifyCanExecuteChangedFor(nameof(StartDeploymentCommand))]
-    private string selectedDriverPackVersion = string.Empty;
-
-    [ObservableProperty]
     [NotifyCanExecuteChangedFor(nameof(StartDeploymentCommand))]
     private string cacheRootPath = WinPeTransientRuntimeRoot;
-
-    [ObservableProperty]
-    [NotifyCanExecuteChangedFor(nameof(StartDeploymentCommand))]
-    private bool autoSelectDriverPackWhenEmpty = true;
 
     [ObservableProperty]
     [NotifyCanExecuteChangedFor(nameof(StartDeploymentCommand))]
@@ -150,29 +117,16 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
 
     [ObservableProperty]
     private string detectedHardwareSummary = "Detecting hardware...";
-    public ObservableCollection<DriverPackCatalogItem> DriverPacks { get; } = [];
-    public ObservableCollection<DriverPackOptionItem> DriverPackOptions { get; } = [];
-    public ObservableCollection<string> DriverPackModelOptions { get; } = [];
-    public ObservableCollection<string> DriverPackVersionOptions { get; } = [];
     public ObservableCollection<TargetDiskInfo> TargetDisks { get; } = [];
     public DeploymentSessionViewModel Session { get; }
     public OperatingSystemCatalogViewModel OperatingSystemCatalog { get; }
+    public DriverPackSelectionViewModel DriverPackSelection { get; }
 
     public DeployThemeMode CurrentTheme => _themeService.CurrentTheme;
     public bool IsDebugSafeMode => DebugSafetyMode.IsEnabled;
     public string EffectiveOsArchitecture => OperatingSystemCatalog.EffectiveOsArchitecture;
     public OperatingSystemCatalogItem? SelectedOperatingSystem => OperatingSystemCatalog.SelectedOperatingSystem;
-    public string DriverPackModeDisplay => SelectedDriverPackOption?.Kind switch
-    {
-        DriverPackSelectionKind.MicrosoftUpdateCatalog => "Microsoft Update Catalog",
-        DriverPackSelectionKind.OemCatalog => "OEM Driver Pack",
-        _ => "None"
-    };
-    public bool IsOemDriverSourceSelected => SelectedDriverPackOption?.Kind == DriverPackSelectionKind.OemCatalog;
-    public bool IsDriverPackModelSelectionEnabled => IsOemDriverSourceSelected && DriverPackModelOptions.Count > 0;
-    public bool IsDriverPackVersionSelectionEnabled => IsDriverPackModelSelectionEnabled && DriverPackVersionOptions.Count > 0;
     public bool IsFirmwareUpdatesOptionEnabled => _detectedHardware?.IsVirtualMachine != true;
-    public string SelectedDriverPackSelectionDisplay => BuildSelectedDriverPackSelectionDisplay();
     public bool HasTargetComputerNameValidationError => !string.IsNullOrWhiteSpace(TargetComputerNameValidationMessage);
     public string VersionDisplay => $"Version: {AppVersion}";
 
@@ -215,7 +169,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
         _hardwareProfileService = hardwareProfileService;
         _offlineWindowsComputerNameService = offlineWindowsComputerNameService;
         _targetDiskService = targetDiskService;
-        _driverPackSelectionService = driverPackSelectionService;
         _logger = logger;
         _dispatcher = Application.Current?.Dispatcher ?? Dispatcher.CurrentDispatcher;
         (DeploymentMode resolvedMode, string? resolvedUsbCacheRuntimeRoot) = ResolveDeploymentRuntimeContext();
@@ -225,6 +178,8 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             _logger,
             Environment.GetEnvironmentVariable("PROCESSOR_ARCHITECTURE") ?? string.Empty);
         OperatingSystemCatalog.StateChanged += OnOperatingSystemCatalogStateChanged;
+        DriverPackSelection = new DriverPackSelectionViewModel(driverPackSelectionService);
+        DriverPackSelection.StateChanged += OnDriverPackSelectionStateChanged;
         Session = new DeploymentSessionViewModel(
             _dispatcher,
             _logger,
@@ -338,15 +293,12 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             RunOnUi(() =>
             {
                 OperatingSystemCatalog.ApplyCatalog(operatingSystems);
-                DriverPacks.Clear();
-                foreach (DriverPackCatalogItem item in driverPacks)
-                {
-                    DriverPacks.Add(item);
-                }
+                DriverPackSelection.ReplaceCatalog(driverPacks);
+                DriverPackSelection.SetOperatingSystemContext(
+                    OperatingSystemCatalog.SelectedOperatingSystem,
+                    OperatingSystemCatalog.EffectiveOsArchitecture);
 
-                RefreshDriverPackOptions();
-
-                Session.SetStatus($"Catalogs loaded: {OperatingSystemCatalog.OperatingSystems.Count} OS entries, {DriverPacks.Count} driver packs.");
+                Session.SetStatus($"Catalogs loaded: {OperatingSystemCatalog.OperatingSystems.Count} OS entries, {DriverPackSelection.DriverPacks.Count} driver packs.");
             });
         }
         catch (Exception ex)
@@ -484,8 +436,8 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             return;
         }
 
-        DriverPackSelectionKind effectiveDriverPackKind = SelectedDriverPackOption?.Kind ?? DriverPackSelectionKind.None;
-        DriverPackCatalogItem? effectiveDriverPack = ResolveEffectiveDriverPackSelection();
+        DriverPackSelectionKind effectiveDriverPackKind = DriverPackSelection.EffectiveSelectionKind;
+        DriverPackCatalogItem? effectiveDriverPack = DriverPackSelection.ResolveEffectiveSelection();
 
         if (effectiveDriverPackKind == DriverPackSelectionKind.OemCatalog &&
             effectiveDriverPack is null)
@@ -611,27 +563,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
         TargetComputerNameValidationMessage = ComputerNameRules.GetValidationMessage(normalized);
     }
 
-    partial void OnSelectedDriverPackOptionChanged(DriverPackOptionItem? value)
-    {
-        if (_isUpdatingDriverPackOptionSelection)
-        {
-            return;
-        }
-
-        _hasUserSelectedDriverPackOption = true;
-        RefreshDriverPackModelAndVersionOptions();
-    }
-
-    partial void OnSelectedDriverPackModelChanged(string value)
-    {
-        RefreshDriverPackVersionOptions();
-    }
-
-    partial void OnAutoSelectDriverPackWhenEmptyChanged(bool value)
-    {
-        RefreshDriverPackOptions();
-    }
-
     partial void OnApplyFirmwareUpdatesChanged(bool value)
     {
         if (_isUpdatingFirmwareOptionSelection)
@@ -660,8 +591,15 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     {
         OnPropertyChanged(nameof(EffectiveOsArchitecture));
         OnPropertyChanged(nameof(SelectedOperatingSystem));
-        RefreshDriverPackOptions();
+        DriverPackSelection.SetOperatingSystemContext(
+            OperatingSystemCatalog.SelectedOperatingSystem,
+            OperatingSystemCatalog.EffectiveOsArchitecture);
         NextWizardStepCommand.NotifyCanExecuteChanged();
+        StartDeploymentCommand.NotifyCanExecuteChanged();
+    }
+
+    private void OnDriverPackSelectionStateChanged(object? sender, EventArgs e)
+    {
         StartDeploymentCommand.NotifyCanExecuteChanged();
     }
 
@@ -744,12 +682,7 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
 
     private bool HasValidDriverPackSelection()
     {
-        if (SelectedDriverPackOption?.Kind != DriverPackSelectionKind.OemCatalog)
-        {
-            return true;
-        }
-
-        return ResolveEffectiveDriverPackSelection() is not null;
+        return DriverPackSelection.HasValidSelection();
     }
 
     private string NormalizeManagedComputerNameValue(string? value)
@@ -866,542 +799,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             : normalized;
     }
 
-    private void RefreshDriverPackOptions()
-    {
-        string previousKey = _hasUserSelectedDriverPackOption
-            ? SelectedDriverPackOption?.Key ?? string.Empty
-            : string.Empty;
-        DriverPackOptionItem[] options = BuildDriverPackOptions();
-
-        _isUpdatingDriverPackOptionSelection = true;
-        try
-        {
-            DriverPackOptions.Clear();
-            foreach (DriverPackOptionItem option in options)
-            {
-                DriverPackOptions.Add(option);
-            }
-
-            DriverPackOptionItem? selected = null;
-            if (!string.IsNullOrWhiteSpace(previousKey))
-            {
-                selected = options.FirstOrDefault(option =>
-                    option.Key.Equals(previousKey, StringComparison.OrdinalIgnoreCase));
-            }
-
-            selected ??= ResolveDefaultDriverPackOption(options);
-            SelectedDriverPackOption = selected;
-        }
-        finally
-        {
-            _isUpdatingDriverPackOptionSelection = false;
-        }
-
-        RefreshDriverPackModelAndVersionOptions();
-    }
-
-    private DriverPackOptionItem[] BuildDriverPackOptions()
-    {
-        return
-        [
-            CreateNoneDriverPackOption(),
-            CreateMicrosoftUpdateCatalogOption(),
-            CreateOemDriverPackOption(DellDriverPackOptionKey, "Dell"),
-            CreateOemDriverPackOption(LenovoDriverPackOptionKey, "Lenovo"),
-            CreateOemDriverPackOption(HpDriverPackOptionKey, "HP"),
-            CreateOemDriverPackOption(MicrosoftOemDriverPackOptionKey, "Microsoft")
-        ];
-    }
-
-    private DriverPackCatalogItem[] BuildSourceDriverPackCandidates()
-    {
-        if (!IsOemDriverSourceSelected)
-        {
-            return [];
-        }
-
-        string sourceManufacturer = ResolveManufacturerFromSourceOptionKey(SelectedDriverPackOption?.Key ?? string.Empty);
-        if (string.IsNullOrWhiteSpace(sourceManufacturer))
-        {
-            return [];
-        }
-
-        return BuildFilteredDriverPackCandidates(forceManufacturer: sourceManufacturer);
-    }
-
-    private void RefreshDriverPackModelAndVersionOptions()
-    {
-        string previousModel = SelectedDriverPackModel;
-
-        DriverPackModelOptions.Clear();
-        DriverPackVersionOptions.Clear();
-        SelectedDriverPackModel = string.Empty;
-        SelectedDriverPackVersion = string.Empty;
-
-        if (!IsOemDriverSourceSelected)
-        {
-            NotifyDriverPackSelectionStateChanged();
-            return;
-        }
-
-        DriverPackCatalogItem[] sourceCandidates = BuildSourceDriverPackCandidates();
-        string[] models = sourceCandidates
-            .SelectMany(GetSelectableModelNames)
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .OrderBy(model => model, StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-
-        foreach (string model in models)
-        {
-            DriverPackModelOptions.Add(model);
-        }
-
-        if (models.Length > 0)
-        {
-            string preferredModel = models.FirstOrDefault(model =>
-                model.Equals(previousModel, StringComparison.OrdinalIgnoreCase))
-                ?? ResolvePreferredModelFromHardware(sourceCandidates, models);
-
-            SelectedDriverPackModel = preferredModel;
-        }
-
-        NotifyDriverPackSelectionStateChanged();
-    }
-
-    private void RefreshDriverPackVersionOptions()
-    {
-        string previousVersion = SelectedDriverPackVersion;
-        DriverPackVersionOptions.Clear();
-        SelectedDriverPackVersion = string.Empty;
-
-        if (!IsOemDriverSourceSelected || string.IsNullOrWhiteSpace(SelectedDriverPackModel))
-        {
-            NotifyDriverPackSelectionStateChanged();
-            return;
-        }
-
-        DriverPackCatalogItem[] modelCandidates = FilterDriverPackCandidatesBySelectedModel(BuildSourceDriverPackCandidates());
-        DriverPackCatalogItem[] orderedCandidates = SortDriverPackCandidates(modelCandidates);
-
-        string[] versions = orderedCandidates
-            .Select(GetDriverPackVersionDisplay)
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-
-        foreach (string version in versions)
-        {
-            DriverPackVersionOptions.Add(version);
-        }
-
-        if (versions.Length > 0)
-        {
-            SelectedDriverPackVersion = versions.FirstOrDefault(version =>
-                version.Equals(previousVersion, StringComparison.OrdinalIgnoreCase))
-                ?? versions[0];
-        }
-
-        NotifyDriverPackSelectionStateChanged();
-    }
-
-    private DriverPackCatalogItem[] FilterDriverPackCandidatesBySelectedModel(IEnumerable<DriverPackCatalogItem> candidates)
-    {
-        if (string.IsNullOrWhiteSpace(SelectedDriverPackModel))
-        {
-            return candidates.ToArray();
-        }
-
-        string selectedModel = SelectedDriverPackModel.Trim();
-        return candidates
-            .Where(item => GetSelectableModelNames(item).Any(model =>
-                model.Equals(selectedModel, StringComparison.OrdinalIgnoreCase)))
-            .ToArray();
-    }
-
-    private string ResolvePreferredModelFromHardware(
-        IReadOnlyList<DriverPackCatalogItem> sourceCandidates,
-        IReadOnlyList<string> modelOptions)
-    {
-        if (modelOptions.Count == 0)
-        {
-            return string.Empty;
-        }
-
-        if (_detectedHardware is null)
-        {
-            return modelOptions[0];
-        }
-
-        string[] hardwareTokens =
-        [
-            _detectedHardware.Model.Trim(),
-            _detectedHardware.Product.Trim()
-        ];
-        hardwareTokens = hardwareTokens
-            .Where(token => !string.IsNullOrWhiteSpace(token))
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-
-        if (hardwareTokens.Length == 0)
-        {
-            return modelOptions[0];
-        }
-
-        string? exactOptionMatch = modelOptions.FirstOrDefault(option =>
-            hardwareTokens.Any(token => option.Equals(token, StringComparison.OrdinalIgnoreCase)));
-        if (!string.IsNullOrWhiteSpace(exactOptionMatch))
-        {
-            return exactOptionMatch;
-        }
-
-        string? containsOptionMatch = modelOptions.FirstOrDefault(option =>
-            hardwareTokens.Any(token => IsFuzzyModelMatch(option, token)));
-        if (!string.IsNullOrWhiteSpace(containsOptionMatch))
-        {
-            return containsOptionMatch;
-        }
-
-        DriverPackCatalogItem? bestPackMatch = sourceCandidates
-            .Where(item => item.ModelNames.Any(modelName =>
-                hardwareTokens.Any(token => IsFuzzyModelMatch(modelName, token))))
-            .OrderByDescending(item => item.ReleaseDate ?? DateTimeOffset.MinValue)
-            .ThenBy(item => item.Name, StringComparer.OrdinalIgnoreCase)
-            .FirstOrDefault();
-
-        if (bestPackMatch is not null)
-        {
-            string? modelFromPack = GetSelectableModelNames(bestPackMatch)
-                .FirstOrDefault(model => modelOptions.Any(option =>
-                    option.Equals(model, StringComparison.OrdinalIgnoreCase)));
-            if (!string.IsNullOrWhiteSpace(modelFromPack))
-            {
-                return modelFromPack;
-            }
-        }
-
-        return modelOptions[0];
-    }
-
-    private DriverPackCatalogItem[] BuildFilteredDriverPackCandidates(string forceManufacturer = "")
-    {
-        IEnumerable<DriverPackCatalogItem> query = DriverPacks;
-
-        string architecture = NormalizeArchitecture(SelectedOperatingSystem?.Architecture ?? EffectiveOsArchitecture);
-        if (!string.IsNullOrWhiteSpace(architecture))
-        {
-            query = query.Where(item => IsArchitectureMatch(architecture, item.OsArchitecture));
-        }
-
-        string selectedOsRelease = SelectedOperatingSystem?.WindowsRelease?.Trim() ?? string.Empty;
-        if (!string.IsNullOrWhiteSpace(selectedOsRelease))
-        {
-            string windowsLabel = $"Windows {selectedOsRelease}";
-            query = query.Where(item =>
-                item.OsName.Contains(windowsLabel, StringComparison.OrdinalIgnoreCase) ||
-                item.OsName.Contains(selectedOsRelease, StringComparison.OrdinalIgnoreCase));
-        }
-
-        if (!string.IsNullOrWhiteSpace(forceManufacturer))
-        {
-            query = query.Where(item =>
-                NormalizeManufacturer(item.Manufacturer).Equals(forceManufacturer, StringComparison.OrdinalIgnoreCase));
-        }
-
-        DriverPackCatalogItem[] baseCandidates = query.ToArray();
-        return SortDriverPackCandidates(baseCandidates);
-    }
-
-    private DriverPackOptionItem? ResolveDefaultDriverPackOption(IReadOnlyList<DriverPackOptionItem> options)
-    {
-        if (options.Count == 0)
-        {
-            return null;
-        }
-
-        if (!AutoSelectDriverPackWhenEmpty)
-        {
-            return options.FirstOrDefault(option => option.Kind == DriverPackSelectionKind.None) ?? options[0];
-        }
-
-        if (_detectedHardware?.IsVirtualMachine == true)
-        {
-            return options.FirstOrDefault(option => option.Kind == DriverPackSelectionKind.MicrosoftUpdateCatalog)
-                   ?? options[0];
-        }
-
-        if (_detectedHardware is not null && SelectedOperatingSystem is not null && DriverPacks.Count > 0)
-        {
-            DriverPackSelectionResult selection = _driverPackSelectionService.SelectBest(
-                DriverPacks.ToArray(),
-                _detectedHardware,
-                SelectedOperatingSystem);
-
-            if (selection.DriverPack is not null)
-            {
-                string selectedKey = ResolveSourceOptionKey(selection.DriverPack.Manufacturer);
-                DriverPackOptionItem? oemMatch = options.FirstOrDefault(option =>
-                    option.Key.Equals(selectedKey, StringComparison.OrdinalIgnoreCase));
-
-                if (oemMatch is not null)
-                {
-                    return oemMatch;
-                }
-            }
-        }
-
-        return options.FirstOrDefault(option => option.Kind == DriverPackSelectionKind.MicrosoftUpdateCatalog)
-               ?? options[0];
-    }
-
-    private static DriverPackOptionItem CreateNoneDriverPackOption()
-    {
-        return new DriverPackOptionItem
-        {
-            Key = NoneDriverPackOptionKey,
-            DisplayName = "None",
-            Kind = DriverPackSelectionKind.None,
-            DriverPack = null
-        };
-    }
-
-    private static DriverPackOptionItem CreateMicrosoftUpdateCatalogOption()
-    {
-        return new DriverPackOptionItem
-        {
-            Key = MicrosoftUpdateCatalogDriverPackOptionKey,
-            DisplayName = "Microsoft Update Catalog",
-            Kind = DriverPackSelectionKind.MicrosoftUpdateCatalog,
-            DriverPack = null
-        };
-    }
-
-    private static DriverPackOptionItem CreateOemDriverPackOption(string key, string displayName)
-    {
-        return new DriverPackOptionItem
-        {
-            Key = key,
-            DisplayName = displayName,
-            Kind = DriverPackSelectionKind.OemCatalog,
-            DriverPack = null
-        };
-    }
-
-    private DriverPackCatalogItem? ResolveEffectiveDriverPackSelection()
-    {
-        DriverPackSelectionKind selectionKind = SelectedDriverPackOption?.Kind ?? DriverPackSelectionKind.None;
-        if (selectionKind != DriverPackSelectionKind.OemCatalog)
-        {
-            return SelectedDriverPackOption?.DriverPack;
-        }
-
-        DriverPackCatalogItem[] sourceCandidates = BuildSourceDriverPackCandidates();
-        if (sourceCandidates.Length == 0)
-        {
-            return null;
-        }
-
-        DriverPackCatalogItem[] modelCandidates = FilterDriverPackCandidatesBySelectedModel(sourceCandidates);
-        if (modelCandidates.Length == 0)
-        {
-            return null;
-        }
-
-        DriverPackCatalogItem[] versionCandidates = string.IsNullOrWhiteSpace(SelectedDriverPackVersion)
-            ? modelCandidates
-            : modelCandidates
-                .Where(item => GetDriverPackVersionDisplay(item).Equals(SelectedDriverPackVersion.Trim(), StringComparison.OrdinalIgnoreCase))
-                .ToArray();
-
-        DriverPackCatalogItem[] finalCandidates = versionCandidates.Length > 0
-            ? versionCandidates
-            : modelCandidates;
-
-        return finalCandidates
-            .OrderByDescending(item => item.ReleaseDate ?? DateTimeOffset.MinValue)
-            .ThenBy(item => item.Name, StringComparer.OrdinalIgnoreCase)
-            .FirstOrDefault();
-    }
-
-    private string BuildSelectedDriverPackSelectionDisplay()
-    {
-        DriverPackSelectionKind selectionKind = SelectedDriverPackOption?.Kind ?? DriverPackSelectionKind.None;
-        if (selectionKind == DriverPackSelectionKind.None)
-        {
-            return "None";
-        }
-
-        if (selectionKind == DriverPackSelectionKind.MicrosoftUpdateCatalog)
-        {
-            return "Microsoft Update Catalog";
-        }
-
-        DriverPackCatalogItem? selectedPack = ResolveEffectiveDriverPackSelection();
-        if (selectedPack is null)
-        {
-            string sourceName = SelectedDriverPackOption?.DisplayName ?? "OEM";
-            return $"{sourceName} | No matching model/version";
-        }
-
-        string modelName = string.IsNullOrWhiteSpace(SelectedDriverPackModel)
-            ? ResolveDriverPackFriendlyName(selectedPack)
-            : SelectedDriverPackModel;
-        string version = GetDriverPackVersionDisplay(selectedPack);
-
-        return $"{selectedPack.Manufacturer} | {modelName} | {version}";
-    }
-
-    private static string ResolveManufacturerFromSourceOptionKey(string optionKey)
-    {
-        return optionKey.Trim().ToLowerInvariant() switch
-        {
-            DellDriverPackOptionKey => "dell",
-            LenovoDriverPackOptionKey => "lenovo",
-            HpDriverPackOptionKey => "hp",
-            MicrosoftOemDriverPackOptionKey => "microsoft",
-            _ => string.Empty
-        };
-    }
-
-    private static string ResolveSourceOptionKey(string manufacturer)
-    {
-        string normalized = NormalizeManufacturer(manufacturer);
-        return normalized switch
-        {
-            "dell" => DellDriverPackOptionKey,
-            "lenovo" => LenovoDriverPackOptionKey,
-            "hp" => HpDriverPackOptionKey,
-            "microsoft" => MicrosoftOemDriverPackOptionKey,
-            _ => string.Empty
-        };
-    }
-
-    private static string[] GetSelectableModelNames(DriverPackCatalogItem driverPack)
-    {
-        string[] models = driverPack.ModelNames
-            .Where(model => !string.IsNullOrWhiteSpace(model))
-            .Select(model => model.Trim())
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-
-        if (models.Length > 0)
-        {
-            return models;
-        }
-
-        string fallback = ResolveDriverPackFriendlyName(driverPack);
-        return string.IsNullOrWhiteSpace(fallback)
-            ? []
-            : [fallback];
-    }
-
-    private static string GetDriverPackVersionDisplay(DriverPackCatalogItem driverPack)
-    {
-        if (!string.IsNullOrWhiteSpace(driverPack.Version))
-        {
-            return driverPack.Version.Trim();
-        }
-
-        if (driverPack.ReleaseDate is not null)
-        {
-            return driverPack.ReleaseDate.Value.ToString("yyyy-MM-dd", CultureInfo.InvariantCulture);
-        }
-
-        if (!string.IsNullOrWhiteSpace(driverPack.PackageId))
-        {
-            return driverPack.PackageId.Trim();
-        }
-
-        if (!string.IsNullOrWhiteSpace(driverPack.FileName))
-        {
-            return Path.GetFileNameWithoutExtension(driverPack.FileName.Trim());
-        }
-
-        return "Unknown";
-    }
-
-    private void NotifyDriverPackSelectionStateChanged()
-    {
-        OnPropertyChanged(nameof(IsOemDriverSourceSelected));
-        OnPropertyChanged(nameof(IsDriverPackModelSelectionEnabled));
-        OnPropertyChanged(nameof(IsDriverPackVersionSelectionEnabled));
-        OnPropertyChanged(nameof(SelectedDriverPackSelectionDisplay));
-        StartDeploymentCommand.NotifyCanExecuteChanged();
-    }
-
-    private static string ResolveDriverPackFriendlyName(DriverPackCatalogItem driverPack)
-    {
-        string[] models = driverPack.ModelNames
-            .Where(model => !string.IsNullOrWhiteSpace(model))
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-
-        if (models.Length > 0)
-        {
-            return models.Length == 1
-                ? models[0]
-                : $"{models[0]} (+{models.Length - 1} models)";
-        }
-
-        if (!LooksLikeArchiveOrInstallerName(driverPack.Name))
-        {
-            return driverPack.Name.Trim();
-        }
-
-        if (!LooksLikeArchiveOrInstallerName(driverPack.PackageId))
-        {
-            return driverPack.PackageId.Trim();
-        }
-
-        string fallback = !string.IsNullOrWhiteSpace(driverPack.FileName)
-            ? driverPack.FileName
-            : driverPack.Name;
-
-        return Path.GetFileNameWithoutExtension(fallback).Trim();
-    }
-
-    private static bool LooksLikeArchiveOrInstallerName(string value)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-        {
-            return false;
-        }
-
-        string extension = Path.GetExtension(value.Trim());
-        if (string.IsNullOrWhiteSpace(extension))
-        {
-            return false;
-        }
-
-        return extension.Equals(".exe", StringComparison.OrdinalIgnoreCase) ||
-               extension.Equals(".cab", StringComparison.OrdinalIgnoreCase) ||
-               extension.Equals(".msi", StringComparison.OrdinalIgnoreCase) ||
-               extension.Equals(".zip", StringComparison.OrdinalIgnoreCase) ||
-               extension.Equals(".7z", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static bool ContainsIgnoreCase(string source, string value)
-    {
-        if (string.IsNullOrWhiteSpace(source) || string.IsNullOrWhiteSpace(value))
-        {
-            return false;
-        }
-
-        return source.Contains(value, StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static bool IsFuzzyModelMatch(string source, string value)
-    {
-        return ContainsIgnoreCase(source, value) || ContainsIgnoreCase(value, source);
-    }
-
-    private static DriverPackCatalogItem[] SortDriverPackCandidates(IEnumerable<DriverPackCatalogItem> candidates)
-    {
-        return candidates
-            .OrderByDescending(item => item.ReleaseDate ?? DateTimeOffset.MinValue)
-            .ThenBy(item => item.Manufacturer, StringComparer.OrdinalIgnoreCase)
-            .ThenBy(item => item.Name, StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-    }
-
     private async Task LoadHardwareProfileAsync()
     {
         try
@@ -1411,10 +808,10 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             {
                 _detectedHardware = profile;
                 OperatingSystemCatalog.SetEffectiveArchitecture(profile.Architecture);
+                DriverPackSelection.SetDetectedHardware(profile);
                 SyncFirmwareOptionFromHardware(profile);
                 DetectedHardwareSummary =
                     $"{profile.DisplayLabel} | TPM: {(profile.IsTpmPresent ? "Yes" : "No")} | Autopilot: {(profile.IsAutopilotCapable ? "Capable" : "Needs checks")} | Power: {(profile.IsOnBattery ? "Battery" : "AC")} | Firmware: {(profile.SystemFirmwareHardwareId.Length > 0 ? "Detected" : "Unavailable")}";
-                RefreshDriverPackOptions();
             });
             _logger.LogInformation("Hardware profile loaded in view model. DisplayLabel={DisplayLabel}", profile.DisplayLabel);
         }
@@ -1658,6 +1055,7 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
         }
 
         OperatingSystemCatalog.StateChanged -= OnOperatingSystemCatalogStateChanged;
+        DriverPackSelection.StateChanged -= OnDriverPackSelectionStateChanged;
         Session.Dispose();
         _isDisposed = true;
     }
